chore(contracts): bench cold start and invocation in big eventbridge handler
